Add files via upload
In this project we used the concepts like File I/O and DSA. The game consists of 10 questions. We found around 250 question and put them in a text file. Each question has been allocated 25 different questions to choose from. for Example for question 1 we have 25 different question in  the text file from which any one question is chosen randomly. This prevents repeating of questions as it should be in a quiz game so other players didn't get the questions. Giving correct answer we can go to next question and wrong answer will end the game. More than one player can play the game and in the last score of all the players is shown. In the end it also shows which player got heighest score.